libaudiohal@aidl: adjust pause and flush for a stream
pause: HAL pause can be achieved in one of the
playing/recording kind of states ACTIVE, TRANSFERRING,
or DRAINING. In other states, pause may be treated as success.

flush: HAL flush can be achieved in one of the flushable
states PAUSED, TRANSFER_PAUSED or DRAIN_PAUSED. In other
states, may be treated as success except for playing/recording
kind of states.

BUG: 356083752
Test: play compress offload playback

CRs-Fixed:3857628

Change-Id: Ia7f3461d1c3aff87ab3398f35a431828d305543f