media: Don't override offload info with default
Remove older customization to align with AOSP, to properly set offload info for the tracks. Latest versions of AOSP code has already a mechanism to properly set offload info, if clients are not setting.

This older customizatin is conflicting with pcm offload features when 24/32bit pcm tracks are played.

CRs-Fixed: 4017540

Change-Id: Ie0709e78a1da7499513c0f0f39f8908e3cb8914e