audiopolicy: Remove customization to internally use direct pcm
Remove the customizations to append direct pcm flags, in
the case when client didn't explicitly request for it.
This is causing a behavior issue at compatability tests.

Note that this change doesn't remove the support of direct
pcm. Any client can request with DIRECT flag to use
direct output.

CRs-Fixed: 4032047

Change-Id: I51d8637d40efeb641186e48a2d1bb3172cfb2e03