audiopolicy: select last connected device from same type of devices
In case of multidevice support for same type , always select the last
connected device of given type.

As DeviceVector is sorted vector and because do_compare is ovewritten to
make sure, that if there are multiple devices of same type are connected
the latest connected will be added first in the list. Use this to return
single device for each type.

CRs-Fixed: 4121273

Change-Id: I39d90fdde933e072423c87d7487df968d361cbc5